ai_researcher/requirements.txt
streamlit==1.42.0
google-generativeai==0.8.4
pypdf==5.3.0
python-pptx==1.0.2
python-dotenv==1.0.1
requests==2.32.3
